backend/analysis_service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn
joblib>=1.3.0
scikit-learn>=1.3.0
numpy<2.0.0
torch>=2.2.0
transformers>=4.42.0
xgboost>=2.0.0
python-dotenv==1.0.0
requests>=2.31.0
